+++ tests/Agent/IntegrationTests/IntegrationTests/CustomInstrumentation/InterfaceDefaultsInstrumentationTests.cs
@@ -8,7 +8,7 @@ using NewRelic.Agent.IntegrationTestHelpers;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.IntegrationTests
+namespace NewRelic.Agent.IntegrationTests.CustomInstrumentation
 {
     [NetCoreTest]
     public class InterfaceDefaultsInstrumentationTests : IClassFixture<RemoteServiceFixtures.AspNetCore3FeaturesFixture>
